+++ app/build.gradle
@@ -4,11 +4,11 @@ android {
     compileSdkVersion 23
     buildToolsVersion "23.0.3"
     defaultConfig {
-        applicationId "com.cetcme.rcldandroidJiangxi"
+        applicationId "com.cetcme.rcldandroidZhejiang"
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 1
-        versionName '1.1'
+        versionName '1.0'
 //        multiDexEnabled true //gradle 优化
     }
     buildTypes {
