+++ app/src/main/java/com/cetcme/rcldandroidZhejiang/Base64Encoder.java
@@ -1,4 +1,4 @@
-package com.cetcme.rcldandroidJiangxi;
+package com.cetcme.rcldandroidZhejiang;
 
 /**
  * Created by qiuhong on 5/14/16.
