+++ app/src/main/java/com/cetcme/rcldandroidZhejiang/ChangeInfoActivity.java
@@ -1,4 +1,4 @@
-package com.cetcme.rcldandroidJiangxi;
+package com.cetcme.rcldandroidZhejiang;
 
 import android.app.Activity;
 import android.content.DialogInterface;
