+++ app/src/main/java/com/cetcme/rcldandroidZhejiang/UpdateAppManager.java
@@ -1,4 +1,4 @@
-package com.cetcme.rcldandroidJiangxi;
+package com.cetcme.rcldandroidZhejiang;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
