+++ app/src/main/java/com/cetcme/rcldandroidZhejiang/settingActivity.java
@@ -1,4 +1,4 @@
-package com.cetcme.rcldandroidJiangxi;
+package com.cetcme.rcldandroidZhejiang;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
