+++ app/src/main/java/com/cetcme/rcldandroidZhejiang/splashScreen.java
@@ -1,4 +1,4 @@
-package com.cetcme.rcldandroidJiangxi;
+package com.cetcme.rcldandroidZhejiang;
 
 import android.content.Intent;
 import android.content.pm.PackageInfo;
