+++ app/src/main/res/layout/activity_alert_dialog.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.AlertDialogActivity"
+    tools:context="com.cetcme.rcldandroidZhejiang.AlertDialogActivity"
     android:background="#ffffff">
 
     <RelativeLayout
