+++ app/src/main/res/layout/activity_index.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.IndexActivity"
+    tools:context="com.cetcme.rcldandroidZhejiang.IndexActivity"
     android:background="#efeff4">
 
     <RelativeLayout
