+++ app/src/main/res/layout/activity_iof_sailor.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.iofSailorActivity">
+    tools:context="com.cetcme.rcldandroidZhejiang.iofSailorActivity">
 
     <ListView
         android:layout_width="fill_parent"
