+++ app/src/main/res/layout/activity_reason.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.ReasonActivity">
+    tools:context="com.cetcme.rcldandroidZhejiang.ReasonActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
