+++ app/src/main/res/layout/activity_route.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.RouteActivity"
+    tools:context="com.cetcme.rcldandroidZhejiang.RouteActivity"
     android:background="#efeff4">
 
     <RelativeLayout
