+++ app/src/main/res/layout/activity_route_display.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.RouteDisplayActivity">
+    tools:context="com.cetcme.rcldandroidZhejiang.RouteDisplayActivity">
 
     <TextView
         android:layout_width="wrap_content"
