+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.cetcme.rcldandroidJiangxi.splashScreen">
+    tools:context="com.cetcme.rcldandroidZhejiang.splashScreen">
 
     <ImageView
         android:layout_width="150dp"
